refactor: consolidate EditTool into SmartEditTool
